rename: add packages
