+++ src/main/java/gift/api/member/dto/MemberRequest.java
@@ -1,6 +1,6 @@
 package gift.api.member.dto;
 
-import gift.api.member.Member;
+import gift.api.member.domain.Member;
 import gift.api.member.enums.Role;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
