+++ src/main/java/gift/api/product/dto/ProductRequest.java
@@ -1,7 +1,7 @@
 package gift.api.product.dto;
 
 import gift.api.category.Category;
-import gift.api.product.Product;
+import gift.api.product.domain.Product;
 import gift.api.product.validator.NoKakao;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
