+++ src/main/java/gift/api/wishlist/dto/WishAddUpdateRequest.java
@@ -1,7 +1,7 @@
 package gift.api.wishlist.dto;
 
-import gift.api.member.Member;
-import gift.api.product.Product;
+import gift.api.member.domain.Member;
+import gift.api.product.domain.Product;
 import gift.api.wishlist.domain.Wish;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
