+++ src/test/java/gift/api/member/MemberRepositoryTest.java
@@ -3,7 +3,9 @@ package gift.api.member;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import gift.api.member.domain.Member;
 import gift.api.member.enums.Role;
+import gift.api.member.repository.MemberRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
