+++ src/test/java/gift/api/option/OptionIntegrationTest.java
@@ -5,8 +5,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import gift.api.category.Category;
 import gift.api.category.CategoryRepository;
 import gift.api.option.domain.Option;
-import gift.api.product.Product;
-import gift.api.product.ProductRepository;
+import gift.api.option.repository.OptionRepository;
+import gift.api.option.service.OptionService;
+import gift.api.product.domain.Product;
+import gift.api.product.repository.ProductRepository;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
