+++ src/test/java/gift/api/option/OptionServiceTest.java
@@ -11,8 +11,10 @@ import gift.api.option.domain.Option;
 import gift.api.option.dto.OptionRequest;
 import gift.api.option.dto.OptionResponse;
 import gift.api.option.exception.InvalidNameException;
-import gift.api.product.Product;
-import gift.api.product.ProductRepository;
+import gift.api.option.repository.OptionRepository;
+import gift.api.option.service.OptionService;
+import gift.api.product.domain.Product;
+import gift.api.product.repository.ProductRepository;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
