+++ src/test/java/gift/api/option/domain/OptionTest.java
@@ -1,10 +1,11 @@
 package gift.api.option.domain;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.mockito.Mockito.mock;
 
 import gift.api.option.exception.InvalidSubtractionException;
-import gift.api.product.Product;
+import gift.api.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
