+++ src/test/java/gift/api/product/ProductRepositoryTest.java
@@ -5,6 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import gift.api.category.Category;
 import gift.api.category.CategoryRepository;
+import gift.api.product.domain.Product;
+import gift.api.product.repository.ProductRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
