+++ src/test/java/gift/api/wishlist/WishTest.java
@@ -3,8 +3,8 @@ package gift.api.wishlist;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.mock;
 
-import gift.api.member.Member;
-import gift.api.product.Product;
+import gift.api.member.domain.Member;
+import gift.api.product.domain.Product;
 import gift.api.wishlist.domain.Wish;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
